requirements.txt
paho-mqtt
pyaudio
SpeechRecognition
mediapipe==0.10.3
numpy==1.25.2
opencv_contrib_python==4.7.0.72
opencv_python==4.8.1.78